pyproject.toml
[project]
name = "django-hans"
version = "0.1.0"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "argon2-cffi==25.1",
  "celery==5.6.2",
  "coverage==7.13.5",
  "dj-rest-auth==7.2",
  "django==6.0.3",
  "django-allauth==65.15",
  "django-anymail==14",
  "django-celery-beat==2.9",
  "django-cors-headers==4.9",
  "django-coverage-plugin==3.2",
  "django-debug-toolbar==6.2",
  "django-environ==0.13",
  "django-extensions==4.1",
  "django-filter==25.2",
  "django-health-check==4.2.1",
  "django-model-utils==5",
  "django-redis==6",
  "django-storages[s3]==1.14.6",
  "django-stubs[compatible-mypy]==6.0.1",
  "djangorestframework==3.17",
  "djangorestframework-simplejwt==5.5.1",
  "djangorestframework-stubs==3.16.8",
  "drf-spectacular==0.29",
  "factory-boy==3.3.3",
  "flower==2.0.1",
  "gunicorn==25.1",
  "hiredis==3.3.1",
  "ipdb==0.13.13",
  "mypy==1.19.1",
  "pandas==3.0.1",
  "pillow==12.1.1",
  "pre-commit==4.5.1",
  "psycopg==3.3.3",
  "pytest==9.0.2",
  "pytest-django==4.12",
  "pytest-sugar==1.1.1",
  "python-slugify==8.0.4",
  "redis==7.3",
  "ruff==0.15.7",
  "sentry-sdk==2.55",
  "uvicorn[standard]==0.42",
  "uvicorn-worker==0.4",
  "watchfiles==1.1.1",
  "werkzeug[watchdog]==3.1.6",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = [
  "*/migrations/*.py",
  "config/urls.py",
  "staticfiles/*",
]
lint.select = [
  "A",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "B",
  "BLE",
  "C4",
  "C90",
  "COM",
  "DJ",
  "DTZ",
  "E",
  "EM",
  "ERA",
  "EXE",
  "F",
  "FA",
  "FBT",
  "FLY",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "N",
  "PD",
  # "NPY",
  # "AIR",
  "PERF",
  "PGH",
  "PIE",
  "PL",
  "PT",
  # "ARG", # Unused function argument
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RSE",
  # "FURB",
  # "LOG",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TC",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  "COM812", # Trailing comma missing - conflicts with formatter
  "EXE002", # File is executable but no shebang is present — WSL file permission artifact
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "S101",   # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "SIM102", # sometimes it's better to nest
]

[tool.pytest]
# ==== pytest ====

ini_options.minversion = "6.0"
ini_options.addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
ini_options.python_files = [
  "tests.py",
  "test_*.py",
]
ini_options.norecursedirs = [ ".git", "node_modules", ".venv" ]

[tool.coverage]
# ==== Coverage ====

run.include = [ "django_hans/**" ]
run.omit = [ "*/migrations/*", "*/tests/*" ]
run.plugins = [ "django_coverage_plugin" ]

# ==== mypy ====
[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
  "mypy_django_plugin.main",
  "mypy_drf_plugin.main",
]
overrides = [
  # Django migrations should not produce any errors:
  { module = "*.migrations.*", ignore_errors = true },
]

[tool.django-stubs]
django_settings_module = "config.settings.test"
